+++ src/components/admin/donations/grid/GridAppbar.tsx
@@ -7,7 +7,7 @@ import { routes } from 'common/routes'
 import { useMutation } from '@tanstack/react-query'
 import { useExportToExcel } from 'service/donation'
 import { AlertStore } from 'stores/AlertStore'
-import { downloadFile } from './../../../common/util/downloadFile'
+import { downloadFile } from '../../../../common/util/downloadFile'
 import { useMemo, useState } from 'react'
 import { useStores } from 'common/hooks/useStores'
 import { debounce } from 'lodash'
